refactor(frontend): consolidate vendor payment hooks into useVendorPaymentsWorkspace
